pyproject.toml
[project]
name = "ai-project-initiator"
version = "1.0.0"
description = "Structured project initiation workflow with AI assistance"
authors = [{name = "jhc"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "defusedxml>=0.7.1",
    "lxml>=4.9.0",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
    "playwright>=1.40.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/jhc/ai-project-initiator"
Documentation = "https://github.com/jhc/ai-project-initiator/blob/main/README.md"
Repository = "https://github.com/jhc/ai-project-initiator"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
